+++ imio/dms/mail/adapters.py
@@ -40,6 +40,7 @@ from imio.dms.mail.utils import get_scan_id
 from imio.dms.mail.utils import highest_review_level
 from imio.dms.mail.utils import is_dv_conv_in_error
 from imio.dms.mail.utils import logger
+from imio.esign.audit import audit as esign_audit
 from imio.esign.utils import add_files_to_session
 from imio.esign.utils import get_file_download_url
 from imio.esign.utils import get_max_download_date
@@ -1145,8 +1146,8 @@ class OMApprovalAdapter(object):
     Annotation structure: metadata + 2D matrix with n signers and m files
 
     ### Metadata
-    self.annot["session_id"] = str
-        The unique id of the internal approval session
+    self.annot["session_ids"] = list[session_id]
+        The list of esign session ids
     self.annot["signers"] = list[n] (userid, name, label)
         The list of signers for the approval process
     self.annot["approvers"] = list[n] tuple(userids)
@@ -1174,7 +1175,7 @@ class OMApprovalAdapter(object):
                 {
                     # Metadata
                     "current_nb": None,
-                    "session_id": None,
+                    "session_ids": PersistentList(),
                     "signers": PersistentList(),
                     "editors": PersistentList(),
                     "approvers": PersistentList(),
@@ -1190,7 +1191,7 @@ class OMApprovalAdapter(object):
         """Reset approval annotation."""
         # Metadata
         self.annot["current_nb"] = None
-        self.annot["session_id"] = None
+        self.annot["session_ids"] = PersistentList()
         self.annot["signers"] = PersistentList()
         self.annot["editors"] = PersistentList()
         self.annot["approvers"] = PersistentList()
@@ -1200,9 +1201,9 @@ class OMApprovalAdapter(object):
         self.annot["approval"] = PersistentList()
 
     @property
-    def session_id(self):
-        """Return the unique session id of the approval process."""
-        return self.annot["session_id"]
+    def session_ids(self):
+        """Return all session ids for this approval."""
+        return self.annot.get("session_ids", PersistentList())
 
     @property
     def files_uids(self):
@@ -1336,6 +1337,12 @@ class OMApprovalAdapter(object):
     def start_approval_process(self):
         """Update the annotation to start the approval process."""
         orig_nb = self.calculate_current_nb()
+        esign_audit(
+            "start_approval_process",
+            "mail={} signers={} approvers={}".format(
+                self.context.UID(), "|".join(self.signers), "|".join([",".join(sublist) for sublist in
+                                                                      self.annot["approvers"]]))
+        )
         if self.approvers:
             # first time transition, set the first level to "proposed"
             if orig_nb is None:
@@ -1388,6 +1395,7 @@ class OMApprovalAdapter(object):
         # Create approvals backup to restore after update
         backup_files_uids = []
         backup_approvals = []
+        existing_signers = self.signers
         for fuid_index, fuid in enumerate(self.files_uids):
             backup_files_uids.append(fuid)
             backup_approvals.append([])
@@ -1467,6 +1475,14 @@ class OMApprovalAdapter(object):
                     )
                 )
 
+        if existing_signers or backup_approvals:
+            esign_audit(
+                "update_signers",
+                "mail={} signers={} approvers={}".format(
+                    self.context.UID(), "|".join(self.signers), "|".join([",".join(sublist) for sublist in
+                                                                          self.annot["approvers"]]))
+            )
+
         for fuid in backup_files_uids:
             self.add_file_to_approval(fuid)
         if api.content.get_state(self.context) == "to_approve":
@@ -1494,6 +1510,7 @@ class OMApprovalAdapter(object):
             return
         self.annot["files"].append(f_uid)
         self.annot["pdf_files"].append(PersistentList())
+        esign_audit("add_file_to_approval", "mail={} file={}".format(self.context.UID(), f_uid))
         for nb in range(len(self.annot["approval"])):
             self.annot["approval"][nb].append(
                 PersistentMapping(
@@ -1517,6 +1534,7 @@ class OMApprovalAdapter(object):
         file_index = self.files_uids.index(f_uid)
         self.annot["files"].pop(file_index)
         self.annot["pdf_files"].pop(file_index)
+        esign_audit("remove_file_from_approval", "mail={} file={}".format(self.context.UID(), f_uid))
         for nb in range(len(self.annot["approval"])):
             self.annot["approval"][nb].pop(file_index)
         self.annot["current_nb"] = self.calculate_current_nb()
@@ -1610,6 +1628,10 @@ class OMApprovalAdapter(object):
         self.annot["approval"][c_a][f_index]["approved_on"] = datetime.datetime.now()
         self.annot["approval"][c_a][f_index]["approved_by"] = userid
         self.annot["current_nb"] = self.calculate_current_nb()
+        esign_audit(
+            "approve_file",
+            "mail={} file={} signer={} approver={}".format(self.context.UID(), f_uid, self.signers[c_a], userid)
+        )
         pc = getToolByName(self.context, "portal_catalog")
         if self.is_file_approved(f_uid):
             afile.approved = True
@@ -1712,7 +1734,10 @@ class OMApprovalAdapter(object):
         self.annot["approval"][nb][f_index]["status"] = "w"
         self.annot["approval"][nb][f_index]["approved_on"] = None
         self.annot["approval"][nb][f_index]["approved_by"] = None
-
+        esign_audit(
+            "unapprove_file",
+            "mail={} file={} signer={}".format(self.context.UID(), f_uid, signer_userid)
+        )
         afile.approved = False
 
         if orig_nb != self.current_nb:
@@ -1847,6 +1872,7 @@ class OMApprovalAdapter(object):
                 attributes={
                     "content_category": orig_fobj.content_category,
                     "scan_id": orig_fobj.scan_id,
+                    "scan_user": hasattr(orig_fobj, "scan_user") and orig_fobj.scan_user or None,
                     "_plone.uuid": new_uid,
                 },
                 renderer=bool(orig_template),
@@ -1870,6 +1896,10 @@ class OMApprovalAdapter(object):
         # we rename the pdf filename to include pdf uid. So after the file is later consumed, we can retrieve object
         pdf_file.file.filename = u"{}__{}.pdf".format(f_title, pdf_uid)
         session_file_uids.append(pdf_uid)
+        esign_audit(
+            "create_pdf_file",
+            "mail={} file={} pdf={}".format(self.context.UID(), f_uid, pdf_uid)
+        )
 
     def add_mail_files_to_session(self):
         """Add mail files to sign session."""
@@ -1934,25 +1964,18 @@ class OMApprovalAdapter(object):
             signers.append((signer, email, name, label))
         watcher_users = api.user.get_users(groupname="esign_watchers")
         watcher_emails = [user.getProperty("email") for user in watcher_users]
-        session_id, _session = add_files_to_session(
-            signers,
-            session_file_uids,
-            bool(self.context.seal),
-            title=_("[ia.docs] Session {sign_id}"),
-            watchers=watcher_emails,
-        )
-        self.annot["session_id"] = session_id
-        session_len = len(session_file_uids)
-        if session_len > 1:
-            return True, _(
-                "${count} files added to session number ${session_id}",
-                mapping={"count": session_len, "session_id": session_id},
-            )
-        else:
-            return True, _(
-                "${count} file added to session number ${session_id}",
-                mapping={"count": session_len, "session_id": session_id},
-            )
+        # Add one file at a time so max_session_size discrimination splits sessions naturally
+        pdf_session_ids = set()
+        for pdf_uid in session_file_uids:
+            sid, _session = add_files_to_session(signers, [pdf_uid], bool(self.context.seal),
+                                                 title=_("[ia.docs] Session {sign_id}"),
+                                                 watchers=watcher_emails)
+            pdf_session_ids.add(sid)
+            if sid not in self.annot["session_ids"]:
+                self.annot["session_ids"].append(sid)
+        return True, _("${count} file(s) added to session(s) ${session_ids}",
+                       mapping={"count": str(len(session_file_uids)),
+                                "session_ids": u", ".join([str(sid) for sid in pdf_session_ids])})
 
 
 class DmsCategorizedObjectInfoAdapter(CategorizedObjectInfoAdapter):